// src/types/Inputs.ts
export type Inputs = {
  date?: string;
  program: string;
  level: string;
  description: string;
  modality: string;
  institution: string;
  period: string;
  durationTime: number;
  typePeriod: string;
  academicPeriods: string;
  studyLeave: string;
  status: string;
  startDate: string;
  degreeDate: string;
  totalCost: number;
  additionalFinancial: boolean;
  percentageSupport: number;
  cashSupport: number;
};